+++ kernel/app/src/js/publish.js
@@ -1,14 +1,17 @@
 (() => {
     const url = new URL(window.location.href);
+    const REG = {
+        url: /^siyuan:\/\/blocks\/(\d{14}\-[0-9a-z]{7})/,
+    };
 
     /* 将块引用转化为超链接 */
-    const a_url = new URL(url);
-    a_url.pathname = "/block";
+    const publish_url = new URL(url);
+    publish_url.pathname = "/block";
     document.querySelectorAll(`#preview span[data-type="block-ref"][data-id]`).forEach(item => {
         const id = item.dataset.id;
         const a = document.createElement("a");
-        a_url.searchParams.set("id", id);
-        a.href = a_url.href;
+        publish_url.searchParams.set("id", id);
+        a.href = publish_url.href;
         // a.target = "_blank";
         item.parentElement.replaceChild(a, item);
         a.appendChild(item);
@@ -16,12 +19,23 @@
 
     /* 将链接转化为超链接 */
     document.querySelectorAll(`#preview span[data-type="a"][data-href]`).forEach(item => {
-        const href = item.dataset.href;
         const a = document.createElement("a");
+        let href = item.dataset.href;
+        if (REG.url.test(href)) { // 思源块超链接转化为站点超链接
+            const id = REG.url.exec(href)[1];
+            publish_url.searchParams.set("id", id);
+            href = publish_url.href;
+        }
         a.href = href;
         // a.target = "_blank";
         item.parentElement.replaceChild(a, item);
         a.appendChild(item);
+    });
+
+    /* 修改块超链接的 href */
+    document.querySelectorAll(`#preview a[href ^="siyuan:"]`).forEach(item => {
+        const href = item.href;
+        item.href;
     });
 
     /* 定位到指定的块并高亮 */